# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api.createSubstitutionRequest:
    properties:
      class_timetable_slot_id:
        type: string
      original_teacher_membership_id:
        type: string
      reason:
        type: string
      status:
        type: string
      substitute_teacher_membership_id:
        type: string
      substitution_date:
        type: string
    type: object
...
  internal_api.updateSubstitutionRequest:
    properties:
      id:
        type: string
      reason:
        type: string
      status:
        type: string
      substitute_teacher_membership_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Platform team
  description: Backend REST API for Somotracker educational dashboard
  license:
    name: MIT
  termsOfService: http://somotracker.local/terms
  title: Somotracker API
  version: 1.0.0
paths:
...
  /api/timetable/substitutions:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: YYYY-MM-DD
        in: query
        name: date_from
        type: string
      - description: YYYY-MM-DD
        in: query
        name: date_to
        type: string
      - description: 'Filter: PENDING, ASSIGNED, COMPLETED, CANCELLED'
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: List timetable substitutions
      tags:
      - Timetable
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.updateSubstitutionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Update timetable substitution
      tags:
      - Timetable
    post:
      consumes:
      - application/json
      parameters:
      - description: Substitution payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.createSubstitutionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Create timetable substitution
      tags:
      - Timetable
  /api/timetable/substitutions/{id}:
    delete:
      parameters:
      - description: Substitution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Delete timetable substitution
      tags:
      - Timetable
    get:
      parameters:
      - description: Substitution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get timetable substitution
      tags:
      - Timetable
...
swagger: "2.0"